pyproject.toml
[project]
name = "agentic-research-assistant"
version = "0.1.0"
description = "A production-grade agentic research assistant with MCP tools and full observability"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "pypdf>=5.4.0",
    "pdfplumber>=0.11.0",
    "docker>=7.1.0",
    "tavily-python>=0.5.0",
    "opentelemetry-sdk>=1.31.0",
    "opentelemetry-exporter-otlp>=1.31.0",
    "opentelemetry-instrumentation-httpx>=0.52b0",
    "structlog>=25.1.0",
    "trafilatura>=2.0.0",
    "rich>=13.9.0",
    "typer>=0.15.0",
]

[project.scripts]
agent = "agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
    "testcontainers[postgres]>=4.9.0",
    "ruff>=0.9.0",
    "mypy>=1.15.0",
    "types-docker>=7.0.0",
    "anyio>=4.8.0",
    "pytest-timeout>=2.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "A", "C4", "PT", "RUF"]
ignore = [
    "ANN101",  # missing type annotation for self
    "ANN102",  # missing type annotation for cls
    "S101",    # use of assert (fine in tests)
    "ANN401",  # dynamically typed Any
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 60

[tool.hatch.build.targets.wheel]
packages = ["agent", "mcp_server", "db"]
